+++ apps/dev-tools/src/Bisect/Bisect.tsx
@@ -1,4 +1,6 @@
 import { useEffect, useState } from 'react'
+import { BisectButton } from './BisectButton'
+import { PrItem } from './PrItem'
 import { prNumbers } from './pr-numbers'
 
 export default function Bisect() {
@@ -27,10 +29,12 @@ export default function Bisect() {
 	return (
 		<div className="bisect__wrapper">
 			<p>Use the preview links to test the target behaviour then mark the PR accordingly.</p>
-			{!bisectStarted && <Button onClick={() => setBisectStarted(true)} text={'Start bisect'} />}
+			{!bisectStarted && (
+				<BisectButton onClick={() => setBisectStarted(true)} text={'Start bisect'} />
+			)}
 			{bisectStarted && (
 				<div className="bisect__button-wrapper">
-					<Button onClick={stopBisect} text={'Stop bisect'} />
+					<BisectButton onClick={stopBisect} text={'Stop bisect'} />
 					{!done && (
 						<>
 							<label htmlFor="show-all">
@@ -57,115 +61,24 @@ export default function Bisect() {
 			)}
 			<ul>
 				{prNumbers.map((prNumber: number, index: number) => {
-					const isCurrent = index === currentIndex
-					const isGood = index === goodPrIndex
-					const isBad = index === badPrIndex
-					let color = 'black'
-					let text = prNumber.toString()
-					if (bisectStarted) {
-						if (isCurrent) {
-							color = '#4465E9'
-						} else if (isGood) {
-							color = '#079268'
-							text += '(✅)'
-						} else if (isBad) {
-							color = '#E03130'
-							text += '(❌)'
-						}
-					}
-					const emphasize = color !== 'black'
-
-					const enableButtons =
-						bisectStarted &&
-						((bothMarked && isCurrent) || (!bothMarked && !isGood && !isBad) || showAll)
-					const foundPr = done && index === badPrIndex
-
 					return (
-						<li key={prNumber}>
-							<div className="bisect__item">
-								<a
-									className="mono"
-									style={{
-										color,
-										fontWeight: emphasize ? 'bold' : 'normal',
-										fontSize: emphasize ? '1.2em' : '1em',
-									}}
-									target="_blank"
-									href={`https://github.com/tldraw/tldraw/pull/${prNumber}`}
-									rel="noreferrer"
-								>
-									{text}
-								</a>
-								{enableButtons && !done && (
-									<>
-										<a
-											className="mono"
-											style={{
-												color,
-												fontWeight: emphasize ? 'bold' : 'normal',
-											}}
-											target="_blank"
-											href={`https://pr-${prNumber}-preview-deploy.tldraw.com/`}
-											rel="noreferrer"
-										>
-											(preview)
-										</a>
-										<div className="bisect__good-bad-wrapper">
-											<Button
-												text="✅"
-												title="Mark as good"
-												onClick={() => setGoodPrIndex(index)}
-												type="emoji"
-												emphasize={emphasize}
-											/>
-											<Button
-												text="❌"
-												title="Mark as bad"
-												onClick={() => setBadPrIndex(index)}
-												type="emoji"
-												emphasize={emphasize}
-											/>
-										</div>
-									</>
-								)}
-								{foundPr && <span>{'⬅️ 👀'}</span>}
-							</div>
-						</li>
+						<PrItem
+							key={prNumber}
+							prNumber={prNumber}
+							index={index}
+							currentIndex={currentIndex}
+							badPrIndex={badPrIndex}
+							goodPrIndex={goodPrIndex}
+							done={done}
+							bothMarked={bothMarked}
+							bisectStarted={bisectStarted}
+							setGoodPrIndex={setGoodPrIndex}
+							setBadPrIndex={setBadPrIndex}
+							showAll={showAll}
+						/>
 					)
 				})}
 			</ul>
 		</div>
-	)
-}
-
-function Button({
-	emphasize = false,
-	text,
-	title,
-	type = 'regular',
-	onClick,
-}: {
-	emphasize?: boolean
-	text: string
-	title?: string
-	type?: 'regular' | 'emoji'
-	onClick(): void
-}) {
-	const regular = type === 'regular'
-	return (
-		<button
-			className="bisect__button"
-			style={{
-				backgroundColor: regular ? '#3182ED' : 'transparent',
-				color: regular ? 'white' : 'black',
-				padding: regular ? '10px' : 0,
-				borderRadius: regular ? '5px' : 0,
-				fontSize: emphasize ? '1.2em' : undefined,
-			}}
-			title={title}
-			onClick={onClick}
-		>
-			{text}
-		</button>
 	)
 }
